# Reduce string allocations by comparing issuers in-place