… test/Microsoft.IdentityModel.Validators.Tests/AadIssuerValidatorTests.cs
… test/Microsoft.IdentityModel.Validators.Tests/AadSigningKeyIssuerValidatorTests.cs