Fixes #2595 

### Benchmarks for changes for AadIssuerValidator 
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3593/23H2/2023Update/SunValley3) (Hyper-V)
AMD EPYC 7763, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.100-preview.3.24204.13
  [Host]    : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT AVX2
  MediumRun : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT AVX2

Job=MediumRun  MaxAbsoluteError=10.0000 ms  IterationCount=15
LaunchCount=4  WarmupCount=10

| Method           | Mean     | Error    | StdDev   | Median   | P90      | P95      | P100     | Gen0   | Allocated |
|----------------- |---------:|---------:|---------:|---------:|---------:|---------:|---------:|-------:|----------:|
| Before | 53.03 ns | 0.482 ns | 1.048 ns | 52.65 ns | 54.76 ns | 55.20 ns | 56.80 ns | 0.0086 |     144 B |
| After  | 20.88 ns | 0.026 ns | 0.059 ns | 20.87 ns | 20.94 ns | 20.97 ns | 21.09 ns |      - |         - |

### Benchmarks for changes for AadTokenValidationParametersExtension 
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3593/23H2/2023Update/SunValley3) (Hyper-V)
AMD EPYC 7763, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.100-preview.3.24204.13
  [Host]    : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT AVX2
  MediumRun : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT AVX2

Job=MediumRun  MaxAbsoluteError=10.0000 ms  IterationCount=15
LaunchCount=4  WarmupCount=10

| Method           | Mean     | Error    | StdDev   | P90      | P95      | P100     | Gen0   | Allocated |
|----------------- |---------:|---------:|---------:|---------:|---------:|---------:|-------:|----------:|
| Before | 86.88 ns | 0.542 ns | 1.202 ns | 88.61 ns | 89.01 ns | 90.18 ns | 0.0162 |     272 B |
| After  | 36.93 ns | 0.345 ns | 0.751 ns | 37.80 ns | 38.10 ns | 40.37 ns | 0.0081 |     136 B |