@kellyyangsong 2024-05-31
> :shipit: There were no tests to update?

There were lots of E2E style tests that covered this.. I could write some additional unit tests just for added comfort. Could never hurt!